Adding tests for controller
Refactors the code to move the model population into the controller.
Adds unit test coverage for the model population logic.

Note: The guides I've found for AngularJS testing say to arrange the tests this
way even though it feels to me like the tests cases are overly broad and share
a lot of state. Implementing it this way to start, then maybe will make them
more granular if we find a better way of testing in Angular.